@@ -1,8 +1,10 @@
 // @ts-check
-// Note: type annotations allow type checking and IDEs autocompletion
+// `@type` JSDoc annotations allow editor autocompletion and type checking
+// (when paired with `@ts-check`).
+// There are various equivalent ways to declare your Docusaurus config.
+// See: https://docusaurus.io/docs/api/docusaurus-config
 
-const lightCodeTheme = require('prism-react-renderer/themes/github');
-const darkCodeTheme = require('prism-react-renderer/themes/dracula');
+import {themes as prismThemes} from 'prism-react-renderer';
 
 /** @type {import('@docusaurus/types').Config} */
 const config = {
@@ -16,6 +18,16 @@ const config = {
   organizationName: 'raamcosta', // Usually your GitHub org/user name.
   projectName: 'compose-destinations-docs', // Usually your repo name.
 
+  // Even if you don't use internationalization, you can use this field to set
+  // useful metadata like html lang. For example, if your site is Chinese, you
+  // may want to replace "en" with "zh-Hans".
+  i18n: {
+    defaultLocale: 'en',
+    locales: ['en'],
+  },
+  markdown: {
+    mermaid: true,
+  },
   presets: [
     [
       'classic',
@@ -23,7 +35,7 @@ const config = {
       ({
         docs: {
           routeBasePath: '/',
-          sidebarPath: require.resolve('./sidebars.js'),
+          sidebarPath: './sidebars.js',
           editUrl: 'https://github.com/raamcosta/compose-destinations-docs/edit/main/docusaurus/',
           versions: {
             current: {
@@ -38,7 +50,7 @@ const config = {
         },
         blog: false,
         theme: {
-          customCss: require.resolve('./src/css/custom.css'),
+          customCss: './src/css/custom.css',
         },
       }),
     ],
@@ -71,11 +83,12 @@ const config = {
       },
       prism: {
         additionalLanguages: ['kotlin', 'groovy'],
-        theme: lightCodeTheme,
-        darkTheme: darkCodeTheme,
+        theme: prismThemes.github,
+        darkTheme: prismThemes.dracula,
       },
     }),
     themes: [
+       '@docusaurus/theme-mermaid',
       [
         require.resolve("@easyops-cn/docusaurus-search-local"),
         {
@@ -92,4 +105,4 @@ const config = {
     ],
 };
 
-module.exports = config;
+export default config;
